# StudentManagementApi/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "StudentManagementApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Departments": {
      "get": {
        "tags": [
          "Departments"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Department"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Departments"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DepartmentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Department"
            }
          }
        }
      }
    },
    "/api/Departments/{departmentId}": {
      "get": {
        "tags": [
          "Departments"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "departmentId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Department"
            }
          }
        }
      }
    },
    "/api/Students": {
      "get": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "type": "string"
          },
          {
            "in": "query",
            "name": "fromDate",
            "type": "string"
          },
          {
            "in": "query",
            "name": "toDate",
            "type": "string"
          },
          {
            "in": "query",
            "name": "order",
            "type": "string",
            "default": ""
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 4
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StudentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        }
      }
    },
    "/api/Students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Student"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Department": {
      "type": "object",
      "properties": {
        "departmentId": {
          "format": "int32",
          "type": "integer"
        },
        "departmentName": {
          "type": "string"
        },
        "students": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Student"
          }
        }
      }
    },
    "DepartmentDTO": {
      "type": "object",
      "properties": {
        "departmentName": {
          "type": "string"
        }
      }
    },
    "Student": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "birthdate": {
          "format": "date",
          "type": "string"
        },
        "gender": {
          "type": "boolean"
        },
        "departmentId": {
          "format": "int32",
          "type": "integer"
        },
        "department": {
          "$ref": "#/definitions/Department"
        }
      }
    },
    "StudentDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "birthdate": {
          "format": "date",
          "type": "string"
        },
        "gender": {
          "type": "boolean"
        },
        "departmentId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}